dashboard/requirements.txt
# Core Dashboard Dependencies
streamlit>=1.28.0
plotly>=5.17.0

# Testing Dependencies
pytest>=7.4.0
hypothesis>=6.92.0
pytest-asyncio>=0.21.0

# Data Processing
pandas>=2.0.0

# Markdown Processing
markdown>=3.5.0

# Date/Time Utilities
python-dateutil>=2.8.0

# YAML Configuration
pyyaml>=6.0.0

# AWS SDK for S3 access
boto3>=1.28.0

